Extracted AppNavigation from MainActivity.kt
