+++ auth/src/main/java/com/Found404/paypro/RegistrationResult.kt
@@ -1,0 +1,9 @@
+package com.Found404.paypro
+
+data class RegistrationResult(
+    val success: Boolean,
+    val message: String,
+    val errorCode: Int? = null,
+    val errorMessage: String? = null,
+    val error: String? = null
+)
